refactors: metrics as module
