+++ aiostatsd/metrics/datadog.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass, field
 from typing import Dict
 
-from .basic_metrics import GaugeMetricBase, Metric, TimingMetricBase
+from .basic import GaugeMetricBase, Metric, TimingMetricBase
 
 
 class DatadogMetricBase(Metric):
