+++ tests/test_metrics.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 
-from aiostatsd.basic_metrics import CounterMetric, GaugeMetric, SetMetric, TimingMetric
+from aiostatsd.metrics.basic import CounterMetric, GaugeMetric, SetMetric, TimingMetric
 
 
 class TestCounterMetric(TestCase):
